# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RAG Chat API",
    "description": "Multimodal RAG Platform API (PostgreSQL 16 + pgvector + Hybrid Search)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_documents_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Tùy chọn ghi đè Workspace ID, mặc định lấy từ active workspace",
              "title": "Workspace Id",
              "nullable": true,
              "type": "string",
              "format": "uuid"
            },
            "description": "Tùy chọn ghi đè Workspace ID, mặc định lấy từ active workspace"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "documents:create"
        ]
      },
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_documents_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Tùy chọn ghi đè Workspace ID, mặc định lấy từ active workspace",
              "title": "Workspace Id",
              "nullable": true,
              "type": "string",
              "format": "uuid"
            },
            "description": "Tùy chọn ghi đè Workspace ID, mặc định lấy từ active workspace"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response List Documents Api V1 Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "documents:read"
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "documents:read"
        ]
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "documents:delete"
        ]
      }
    },
    "/documents": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_documents_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Tùy chọn ghi đè Workspace ID, mặc định lấy từ active workspace",
              "title": "Workspace Id",
              "nullable": true,
              "type": "string",
              "format": "uuid"
            },
            "description": "Tùy chọn ghi đè Workspace ID, mặc định lấy từ active workspace"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "documents:create"
        ]
      },
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_documents_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "description": "Tùy chọn ghi đè Workspace ID, mặc định lấy từ active workspace",
              "title": "Workspace Id",
              "nullable": true,
              "type": "string",
              "format": "uuid"
            },
            "description": "Tùy chọn ghi đè Workspace ID, mặc định lấy từ active workspace"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response List Documents Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "documents:read"
        ]
      }
    },
    "/documents/{document_id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "documents:read"
        ]
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "session": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-authentication-required": true,
        "x-permission-mode": "all",
        "x-required-permissions": [
          "documents:delete"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File",
            "format": "binary"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_post"
      },
      "Body_upload_document_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "title": "File",
            "format": "binary"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_documents_post"
      },
...
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_code": {
            "title": "Error Code",
            "nullable": true,
            "type": "string"
          },
          "error_message": {
            "title": "Error Message",
            "nullable": true,
            "type": "string"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count",
            "default": 0
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/IngestionJobResponse"
            },
            "type": "array",
            "title": "Jobs"
          },
          "created_at": {
            "title": "Created At",
            "nullable": true,
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated At",
            "nullable": true,
            "type": "string",
            "format": "date-time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "filename",
          "content_hash",
          "mime_type",
          "file_size",
          "status"
        ],
        "title": "DocumentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestionJobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "parser_name": {
            "type": "string",
            "title": "Parser Name"
          },
          "chunker_name": {
            "type": "string",
            "title": "Chunker Name"
          },
          "elapsed_seconds": {
            "title": "Elapsed Seconds",
            "nullable": true,
            "type": "number"
          },
          "started_at": {
            "title": "Started At",
            "nullable": true,
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "title": "Completed At",
            "nullable": true,
            "type": "string",
            "format": "date-time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "status",
          "retry_count",
          "parser_name",
          "chunker_name"
        ],
        "title": "IngestionJobResponse"
      },
...
      "UploadDocumentResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentResponse"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Tài liệu đã được tải lên và đưa vào hàng đợi xử lý."
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "UploadDocumentResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token",
        "description": "Session token stored in HTTP-only cookie"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token",
        "description": "Session token cookie matching ViShop's @ApiCookieAuth('session')"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token passed in Authorization: Bearer <token>"
      }
    }
  }
}